Fixed buggy status bubbles.
Issue was shared state inside of the system instead of keeping the
state inside of the components

+++ app/src/entity/util/position.ts
@@ -86,6 +86,9 @@ export class PositionUtil extends BaseUtil {
 	): number {
 		const entities = this.getEntitiesWithComponentInTile(tile, componentType);
 		const entityToDestroy = entities[0];
+		if (!entityToDestroy) {
+			return null;
+		}
 		this.entityManager.destroyEntity(entityToDestroy);
 		return entityToDestroy;
 	}
